Zero allocations for SortTags (#68)
This was causing a large number of allocations since it needs to be called for every stats call, which can be VERY frequent. This patch also cuts the cost of SortTags in half. It falls back to using the library sort function over 20 items, because that's where the insertion sort begins to slow down.

Before:
BenchmarkSortTags-2      5000000               337 ns/op              32 B/op          1 allocs/op

After:
BenchmarkSortTags-2     10000000               198 ns/op               0 B/op          0 allocs/op