procstats: use vendor repo for taskstats
With the merge of https://github.com/mdlayher/taskstats/pull/6, so we
can now use the vendor's GitHub repo as the canonical one for the
taskstats library again -- no need to rely on segmentio's fork.

Also, the test code only practically works on Linux, so rename its
file so it's only built there.