move config to responsewriter
responsewriter is also responsible for sending the metrics. So it needs access to the configuration. We could put the configuration to the request writer and do the application there, but if we add configuratinos in the future, it would be nice that they are all applied at the handler level (instead being abstracted into the logic of the request/response writer code)