Rename Go version tag from `version` to `go_version` (#122)
Most of our services will set a version tag that represents the version of their service. However, the Go stats reported by the `procstats` package will overwrite this tag with the current Go runtime version.

This becomes a bigger issue when scraping these stats via Prometheus, since it will consider a sample as invalid if labels are not unique:

![image](https://user-images.githubusercontent.com/2907397/89065716-3345d580-d33a-11ea-93e7-ab4833388980.png)

This PR makes a breaking change to resolve this issue, by renaming the tag that specifics the Go runtime version from `version` to `go_version`.