procstats: bump taskstats and fix file handle leak
The taskstats file handle is not being closed and thus it is possible for
file descriptors to pile up forever. Pull in a much newer version and
fix this problem.

In addition, I don't think the naked returns or check helper are doing any
favors for readability and have removed them in this block in favor of
switch error handling.

Signed-off-by: Matt Layher <mdlayher@gmail.com>