datadog: sanitize metric, field, and tag names
Previously if a metric name contained an invalid character - a colon,
say - we would send it on to the datadog agent. Datadog will reject
metrics containing invalid names, which can lead to problems with the
datadog agent, and/or Datadog.

In addition, truncate metrics that are too long; Datadog will reject
them anyway.