all: bring in objconv repository
We use it as a dependency, but the original repository has been
archived. If we depend on it, we should maintain it.

Bring in all of the code and also update pointers to point at our
version of the code. The tests fail, at the moment.

My hope is that we can prune down the amount of code we need to
maintain, but I started trying to port 'only the code we own' and
ended up porting most of the API, so it was easier to port everything.